@@ -1,128 +1,107 @@
-import 'dart:io';
-import 'dart:math';
-import 'package:path_provider/path_provider.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-import 'package:share_plus/share_plus.dart';
 import 'package:vs_story_designer/vs_story_designer.dart';
-import 'dart:ui' as ui;
 
 void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: 'Flutter story designer Demo',
+      title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        // This is the theme of your application.
+        //
+        // TRY THIS: Try running your application with "flutter run". You'll see
+        // the application has a purple toolbar. Then, without quitting the app,
+        // try changing the seedColor in the colorScheme below to Colors.green
+        // and then invoke "hot reload" (save your changes or press the "hot
+        // reload" button in a Flutter-supported IDE, or press "r" if you used
+        // the command line to start the app).
+        //
+        // Notice that the counter didn't reset back to zero; the application
+        // state is not lost during the reload. To reset the state, use hot
+        // restart instead.
+        //
+        // This works for code too, not just values: Most code changes can be
+        // tested with just a hot reload.
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
       ),
-      home: const Example(),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
-class Example extends StatefulWidget {
-  const Example({Key? key}) : super(key: key);
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
+
+  final String title;
 
   @override
-  State<Example> createState() => _ExampleState();
+  State<MyHomePage> createState() => _MyHomePageState();
 }
 
-class _ExampleState extends State<Example> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        backgroundColor: Colors.white,
-        resizeToAvoidBottomInset: false,
-        body: RepaintBoundary(
-          key: _globalKey,
-          child: Container(
-            color: Colors.white,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const Text(
-                  'Welcome To Story Designer',
-                  textAlign: TextAlign.center,
-                  style: TextStyle(fontSize: 35, fontWeight: FontWeight.bold),
-                ),
-                const SizedBox(height: 10),
-                const Text(
-                  'All New Way To Explore Story Designer',
-                  style: TextStyle(fontSize: 18),
-                ),
-                const SizedBox(height: 50),
-                ElevatedButton(
-                  onPressed: () async {
-                    String? mediaPath = await _prepareImage();
-                    Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => VSStoryDesigner(
-                                  centerText: "Start Creating Your Story",
-                                  // fontFamilyList: const [
-                                  //   FontType.abrilFatface,
-                                  //   FontType.alegreya,
-                                  //   FontType.typewriter
-                                  // ],
-                                  // middleBottomWidget: const SizedBox(),
-                                  themeType: ThemeType
-                                      .light, // OPTIONAL, Default ThemeType.dark
-                                  galleryThumbnailQuality: 250,
-                                  onDone: (uri) {
-                                    debugPrint(uri);
-                                    SharePlus.instance.share(ShareParams(
-                                        files: [XFile(File(uri).path)]));
-                                  },
-                                  mediaPath: mediaPath,
-                                )));
-                  },
-                  child: const Text('Create',
-                      style:
-                          TextStyle(fontSize: 30, fontWeight: FontWeight.w500)),
-                  style: ElevatedButton.styleFrom(
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 20, vertical: 10),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(10),
-                      )),
-                ),
-              ],
+class _MyHomePageState extends State<MyHomePage> {
+
+  void _showStoryDesigner() {
+    Navigator.of(context).push(
+      MaterialPageRoute(
+        builder: (context) => VSStoryDesigner(
+          galleryThumbnailQuality: 200,
+          onDoneButtonStyle: const Padding(
+            padding: EdgeInsets.only(right: 10),
+            child: Icon(
+              Icons.done,
+              size: 40,
+              color: Colors.white,
             ),
           ),
-        ));
+          onDone: (String localLocation) async {
+            Navigator.of(context).pop();
+          },
+          centerText: 'TESTT',
+        ),
+      ),
+    );
   }
 
-  final GlobalKey _globalKey = GlobalKey();
-
-  Future<String?> _prepareImage() async {
-    ByteData? byteData;
-
-    try {
-      RenderRepaintBoundary? boundary = _globalKey.currentContext
-          ?.findRenderObject() as RenderRepaintBoundary?;
-
-      ui.Image? image = await boundary?.toImage(pixelRatio: 4);
-      byteData = await image?.toByteData(format: ui.ImageByteFormat.png);
-      Uint8List bytes = byteData!.buffer.asUint8List();
-
-      final directory = (await getTemporaryDirectory()).path;
-      String imgPath = '$directory/${Random().nextInt(999999)}.jpg';
-      File imgFile = File(imgPath);
-      await imgFile.writeAsBytes(bytes);
-      // Uint8List pngBytes = byteData.buffer.asUint8List();
-      return imgFile.path;
-    } catch (e) {
-      return null;
-    }
+  @override
+  Widget build(BuildContext context) {
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
+    return Scaffold(
+      appBar: AppBar(
+        // TRY THIS: Try changing the color here to a specific color (to
+        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
+        // change color while the other colors stay the same.
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
+        title: Text(widget.title),
+      ),
+      body: Center(),
+      floatingActionButton: FloatingActionButton.extended(
+        onPressed: _showStoryDesigner,
+        tooltip: 'Open Story Designer',
+        label: const Text('Open Story Designer'),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
   }
 }
