@@ -10,6 +10,7 @@ import 'package:vs_story_designer/src/domain/sevices/save_as_image.dart';
 import 'package:vs_story_designer/src/presentation/utils/constants/item_type.dart';
 import 'package:vs_story_designer/src/presentation/utils/constants/text_animation_type.dart';
 import 'package:vs_story_designer/src/presentation/utils/modal_sheets.dart';
+import 'package:vs_story_designer/src/domain/models/vs_story_designer_config.dart';
 import 'package:vs_story_designer/src/presentation/widgets/animated_onTap_button.dart';
 import 'package:vs_story_designer/src/presentation/widgets/tool_button.dart';
 
@@ -17,10 +18,12 @@ class TopTools extends StatefulWidget {
   final GlobalKey contentKey;
   final BuildContext context;
   final Function? renderWidget;
+  final VSStoryDesignerConfig config;
   const TopTools(
       {super.key,
       required this.contentKey,
       required this.context,
+      required this.config,
       this.renderWidget});
 
   @override
@@ -49,7 +52,8 @@ class _TopToolsState extends State<TopTools> {
                       exitDialog(
                               context: widget.context,
                               contentKey: widget.contentKey,
-                              themeType: controlNotifier.themeType)
+                              themeType: controlNotifier.themeType,
+                              config: widget.config)
                           .then((res) {
                         if (res) Navigator.pop(context);
                       });
@@ -181,14 +185,13 @@ class _TopToolsState extends State<TopTools> {
                               saveToGallery: true,
                               fileName: controlNotifier.folderName);
                           if (response) {
-                            Fluttertoast.showToast(msg: 'Successfully saved');
+                            Fluttertoast.showToast(msg: widget.config.successSaveMessage);
                           } else {}
                         }
                         // ignore: use_build_context_synchronously
                         Navigator.of(context, rootNavigator: true).pop();
                       } else {
-                        Fluttertoast.showToast(
-                            msg: 'Design something to save image');
+                        Fluttertoast.showToast(msg: widget.config.emptySaveErrorMessage);
                       }
 
                       setState(() {
