@@ -15,6 +15,7 @@ import 'package:vs_story_designer/src/presentation/utils/Extensions/hexColor.dar
 // import 'package:vs_story_designer/src/presentation/utils/constants/item_type.dart';
 import 'package:vs_story_designer/src/presentation/widgets/animated_onTap_button.dart';
 import 'package:vs_story_designer/vs_story_designer.dart';
+import 'package:vs_story_designer/src/domain/models/vs_story_designer_config.dart';
 
 /// create item of type GIF
 // Future createGiphyItem(
@@ -45,7 +46,8 @@ import 'package:vs_story_designer/vs_story_designer.dart';
 Future<bool> exitDialog(
     {required context,
     required contentKey,
-    required ThemeType themeType}) async {
+    required ThemeType themeType,
+    required VSStoryDesignerConfig config}) async {
   return (await showDialog(
         context: context,
         // barrierColor:
@@ -58,29 +60,29 @@ Future<bool> exitDialog(
           insetAnimationCurve: Curves.ease,
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 15),
-            child: Container(
-              padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-              alignment: Alignment.center,
-              height: 250,
-              decoration: BoxDecoration(
-                  shape: BoxShape.rectangle,
-                  color: themeType == ThemeType.light
-                      ? Colors.white
-                      : HexColor.fromHex('#262626'),
-                  borderRadius: BorderRadius.circular(20),
-                  boxShadow: [
-                    BoxShadow(
-                        color: themeType == ThemeType.light
-                            ? Colors.black
-                            : Colors.white10,
-                        offset: const Offset(0, 1),
-                        blurRadius: 4),
-                  ]),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: <Widget>[
+            child: IntrinsicHeight(
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
+                alignment: Alignment.center,
+                decoration: BoxDecoration(
+                    shape: BoxShape.rectangle,
+                    color: themeType == ThemeType.light
+                        ? Colors.white
+                        : HexColor.fromHex('#262626'),
+                    borderRadius: BorderRadius.circular(20),
+                    boxShadow: [
+                      BoxShadow(
+                          color: themeType == ThemeType.light
+                              ? Colors.black
+                              : Colors.white10,
+                          offset: const Offset(0, 1),
+                          blurRadius: 4),
+                    ]),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: <Widget>[
                   Text(
-                    'Discard Edits?',
+                    config.exitDialogTitle,
                     style: TextStyle(
                         fontSize: 22,
                         fontWeight: FontWeight.w600,
@@ -93,7 +95,7 @@ Future<bool> exitDialog(
                     height: 20,
                   ),
                   Text(
-                    "If you go back now, you'll lose all the edits you've made.",
+                    config.exitDialogMessage,
                     style: TextStyle(
                         fontSize: 15,
                         fontWeight: FontWeight.w400,
@@ -109,7 +111,7 @@ Future<bool> exitDialog(
                           ? Colors.black26
                           : Colors.white10),
                   const SizedBox(height: 10),
-
+                
                   /// discard
                   AnimatedOnTapButton(
                     onTap: () async {
@@ -120,7 +122,7 @@ Future<bool> exitDialog(
                       mainAxisAlignment: MainAxisAlignment.center,
                       children: [
                         Text(
-                          'Discard',
+                          config.exitDialogDiscardButton,
                           style: TextStyle(
                               fontSize: 16,
                               color: Colors.redAccent.shade200,
@@ -137,7 +139,7 @@ Future<bool> exitDialog(
                   //     color: Colors.white10,
                   //   ),
                   // ),
-
+                
                   /// save and exit
                   // AnimatedOnTapButton(
                   //   onTap: () async {
@@ -179,7 +181,7 @@ Future<bool> exitDialog(
                           ? Colors.black26
                           : Colors.white10),
                   const SizedBox(height: 10),
-
+                
                   ///cancel
                   AnimatedOnTapButton(
                     onTap: () {
@@ -189,7 +191,7 @@ Future<bool> exitDialog(
                       mainAxisAlignment: MainAxisAlignment.center,
                       children: [
                         Text(
-                          'Cancel',
+                          config.exitDialogCancelButton,
                           style: TextStyle(
                               fontSize: 16,
                               color: themeType == ThemeType.light
@@ -202,7 +204,9 @@ Future<bool> exitDialog(
                       ],
                     ),
                   ),
-                ],
+                  const SizedBox(height: 10),
+                  ],
+                ),
               ),
             ),
           ),
